Added `.long()` to two identical lines of code in `dyrep.py` and `tgn.py` files.

+++ examples/linkproppred/periodicity_det/tgn.py
@@ -140,7 +140,7 @@ def test(loader, neg_sampler, split_mode):
                     axis=0,
                 ),
                 device=device,
-            )
+            ).long()
 
             n_id = torch.cat([src, dst]).unique()
             n_id, edge_index, e_id = neighbor_loader(n_id)
